+++ prop/formula.test.ts
@@ -1,5 +1,5 @@
 import { assertEquals } from "std/testing/asserts.ts";
-import { parseE1, parseE2, parseE3, parseE4, parseFormula, parsePr } from "./parser.ts";
+import { parseE1, parseE2, parseE3, parseE4, parseFormula, parsePr } from "./formula.ts";
 import { PropFormula } from "./types.ts";
 
 Deno.test("parseE1:P↔Q", () => {
